requirements.txt
Django>=4.2.0
django-rest-framework>=3.14.0
yfinance>=0.2.28
newsapi-python>=0.2.7
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
requests>=2.31.0